requirements.txt
annotated-types==0.7.0
exceptiongroup==1.2.2
iniconfig==2.0.0
mysql-connector-python==9.1.0
mysql-replication==1.0.9
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic-core==2.27.2
pymysql==1.1.1
pytest==8.3.5
python-dotenv==1.0.1
ruff==0.9.10
tomli==2.2.1
typing-extensions==4.12.2
